+++ Feedback/.gitignore
@@ -1,5 +1,5 @@
 # sqlite db
-feedback.db
+*/feedback.db
 
 # Logs
 logs
