--- Feedback/index.js
@@ -1,4 +1,0 @@
-const sqlite3 = require('sqlite3').verbose();
-const backend = require('./backend')
-	
-backend.initialSetup();
