+++ Feedback/package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"No test specified. Its not a bug, its a feature\" && exit 0",
-    "start": "node index.js"
+    "start": "node ./src/index.js"
   },
   "author": "",
   "license": "ISC",
