+++ docker-compose.yml
@@ -2,6 +2,9 @@ version: '3'
 services:
   feedback-service:
     build: ./Feedback
+    volumes:
+      - './Feedback/db:/home/node/app/db'
+      - './Feedback/src:/home/node/app/src'
     expose:
       - "5009"
   food-service:
